63-theme: some clean up, added withStyles story.
