+++ packages/components/src/styles/StylesProvider.tsx
@@ -1,20 +1,8 @@
 import { createContext, useContext, useState, useEffect } from 'react'
 import type { ReactNode } from 'react'
-import { config } from '../../tailwind.config'
+import type { TableStyles } from './'
 import { mergeCustomStyles } from '@stcland/utils'
-
-export type TableStyles = {
-  root: string
-  table: string
-  header: string
-  headerRow: string
-  headerCell: string
-  body: string
-  cell: string
-  row: string
-  selectedRow: string
-  subRow: string
-}
+import { defaultStyles } from './defaults'
 
 export type StclandStyles = {
   table?: Partial<TableStyles>
@@ -22,41 +10,22 @@ export type StclandStyles = {
 
 export type StylesContextType = {
   styles: StclandStyles | null
-  loading: boolean
 }
 
-// Default styles
-const defaultStyles: StclandStyles = {
-  table: {
-    root: 'border border-gray-825 bg-gray-825 px-4 pt-8 pb-14',
-    table: 'table-fixed',
-    header: 'text-s text-gray-100 text-left',
-    headerRow: '',
-    headerCell: 'font-medium pl-6 pb-4',
-    body: 'text-s text-gray-400',
-    row: 'border-t last:border-b border-primary-dark',
-    cell: 'px-6 py-6',
-    selectedRow: 'bg-gray-500',
-    subRow: 'text-secondary-main border-none'
-  }
-}
 
 const StylesContext = createContext<StylesContextType | null>(null)
 export const StcStylesProvider = ({ children, customStyles }: { children: ReactNode, customStyles?: StclandStyles }) => {
   const [styles, setStyles] = useState<StclandStyles>(defaultStyles)
-  const [loading, setLoading] = useState(true)
 
   useEffect(() => {
     const mergedStyles = {
       table: mergeCustomStyles(defaultStyles.table || {}, customStyles?.table || {})
     }
-    console.log('MERGED STYLES:', mergedStyles)
     setStyles(mergedStyles)
-    setLoading(false) // Styles have been loaded
   }, [customStyles])
 
   return (
-    <StylesContext.Provider value={{ styles, loading }}>
+    <StylesContext.Provider value={{ styles }}>
       {children}
     </StylesContext.Provider>
   )
@@ -78,5 +47,3 @@ export const getStyles = (customStyles?: StclandStyles) => {
     table: mergeCustomStyles(defaultStyles.table || {}, customStyles?.table),
   }
 }
-
-export { config as themeConfig }
