+++ packages/components/src/styles/defaults.ts
@@ -1,0 +1,18 @@
+import type { StclandStyles } from "."
+
+// Default styles
+export const defaultStyles: StclandStyles = {
+    table: {
+      root: 'border border-gray-825 bg-gray-825 px-4 pt-8 pb-14',
+      table: 'table-fixed',
+      header: 'text-s text-gray-100 text-left',
+      headerRow: '',
+      headerCell: 'font-medium pl-6 pb-4',
+      body: 'text-s text-gray-400',
+      row: 'border-t last:border-b border-primary-dark',
+      cell: 'px-6 py-6',
+      selectedRow: 'bg-gray-500',
+      subRow: 'text-secondary-main border-none'
+    }
+  }
+
