+++ packages/components/src/styles/index.ts
@@ -1,6 +1,8 @@
 
 import '../index.css'
 
-export { StcStylesProvider, useStyles, getStyles, themeConfig } from './StylesProvider'
+export { StcStylesProvider, useStyles, getStyles } from './StylesProvider'
 
-export type { StclandStyles, TableStyles } from './StylesProvider'
+export type { StclandStyles } from './StylesProvider'
+
+export type { TableStyles } from './table'
