+++ packages/components/src/styles/table.ts
@@ -1,0 +1,13 @@
+export type TableStyles = {
+    root: string
+    table: string
+    header: string
+    headerRow: string
+    headerCell: string
+    body: string
+    cell: string
+    row: string
+    selectedRow: string
+    subRow: string
+  }
+
