+++ packages/components/src/table/withStyles/Table.tsx
@@ -3,13 +3,12 @@ import Table  from '../Table'
 import { useStyles } from '../../styles/StylesProvider'
 
 //*****************************************************************************
-  // Components
+// Components
 //*****************************************************************************
 
 const TableWithStyles = ( props: TableProps) => {
+    
   const { styles }  = useStyles()
-
-  console.log('IN TABLE WRAPPER------------', styles)
 
   return (
     <Table {...props} customStyles={styles?.table}/>
