+++ packages/storybook/stories/Table.mdx
@@ -55,6 +55,19 @@ The row expanded functionality allows users to view additional details within a 
 
 ---
 
+
+## WithStyles Wrapper
+
+When a component is exported from the `@stcland/components/withStyles` module and wrapped in the `StcStylesProvider` it allow 
+users to access styled versions of the core components. These components enhance the base functionality with added styling without
+the need to pass a `customStyles` property as the **With Custom Styles** example above.
+
+
+<Canvas of={TableStories.WithStyles} />
+
+---
+
+
 ## Conclusion
 
 These stories provide an overview of the different ways the table component can be used, styled, and enhanced with interaction. You can further customize these stories or combine features to create more complex tables for your use case.
